+++ src/tracers/mod.rs
@@ -1,3 +1,5 @@
 mod file;
+mod null;
 
 pub use self::file::FileTracer;
+pub use self::null::NullTracer;