+++ src/utils/mod.rs
@@ -1,0 +1,5 @@
+mod global_tracer;
+mod reporter;
+
+pub use self::global_tracer::GlobalTracer;
+pub use self::reporter::ReporterThread;